+++ NEWS.md
@@ -1,5 +1,11 @@
 
 
+# fastglmm 0.3.4
+ - Nov 17, 2025
+ - update `varpart()` for NB models
+  - `fastglmm()` and `glm.nb()`
+  - handle singular models for NB fit
+
 # fastglmm 0.3.3
  - Oct 17, 2025
  - when model fit fails return NaN values